Refactoring to use better naming and folder structure
