+++ src/main/java/com/khanivorous/studentservice/student/StudentRepository.java
@@ -1,5 +1,6 @@
-package com.khanivorous.studentservice;
+package com.khanivorous.studentservice.student;
 
+import com.khanivorous.studentservice.student.Student;
 import org.springframework.data.repository.CrudRepository;
 
 public interface StudentRepository extends CrudRepository<Student, Integer> {
